+++ packages/eslint-plugin-fiori-tools/src/project-context/linker/fe-v4.ts
@@ -45,6 +45,7 @@ export interface AnnotationBasedNode<
     Children = never
 > extends ConfigurationBase<T['type'], Configuration> {
     annotation?: T;
+    label?: string;
     children: Children[];
 }