+++ packages/odata-service-inquirer/src/prompts/datasources/sap-system/service-selection/questions.ts
@@ -89,6 +89,7 @@ export function getSystemServiceQuestion(
             mandatory: true,
             applyDefaultWhenDirty: true
         },
+        showOutputTabLink: 'validationMessageOverflow',
         source: (prevAnswers: unknown, input: string) => searchChoices(input, serviceChoices as ListChoiceOptions[]),
         // SystemSelectionAnswers should not be needed here in the interest of keeping these prompts decoupled but TelemetryHelper is used here and it requires the previously selected destination
         choices: async (answers: SystemSelectionAnswers) => {