+++ packages/odata-service-inquirer/test/unit/prompts/sap-system/abap-on-btp/questions.test.ts
@@ -198,6 +198,7 @@ describe('questions', () => {
                 },
                 "message": [Function],
                 "name": "abapOnBtp:serviceSelection",
+                "showOutputTabLink": "validationMessageOverflow",
                 "source": [Function],
                 "type": "list",
                 "validate": [Function],