+++ packages/odata-service-inquirer/test/unit/prompts/service-url/questions.test.ts
@@ -77,6 +77,7 @@ describe('Service URL prompts', () => {
                 },
                 "message": "OData service URL",
                 "name": "serviceUrl",
+                "showOutputTabLink": "validationMessageOverflow",
                 "type": "input",
                 "validate": [Function],
               },