+++ packages/ui-components/test/unit/components/UICombobox.test.tsx
@@ -79,6 +79,9 @@ describe('<UIComboBox />', () => {
                 "stylesheetKey": "ComboBox",
                 "window": [Window],
               },
+              "callout": Object {
+                "boxShadow": "var(--ui-box-shadow-small)",
+              },
               "errorMessage": Array [
                 Object {
                   "backgroundColor": "var(--vscode-inputValidation-errorBackground)",
@@ -119,6 +122,9 @@ describe('<UIComboBox />', () => {
                 "inShadow": false,
                 "stylesheetKey": "ComboBox",
                 "window": [Window],
+              },
+              "callout": Object {
+                "boxShadow": "var(--ui-box-shadow-small)",
               },
               "errorMessage": Array [
                 Object {