+++ packages/ui-components/test/unit/components/UIDropdown.test.tsx
@@ -51,6 +51,9 @@ describe('<UIDropdown />', () => {
             {},
             `
             Object {
+              "callout": Object {
+                "boxShadow": "var(--ui-box-shadow-small)",
+              },
               "errorMessage": Array [
                 Object {
                   "backgroundColor": "var(--vscode-inputValidation-errorBackground)",
@@ -84,6 +87,9 @@ describe('<UIDropdown />', () => {
         const styles = (wrapper.find(Dropdown).props().styles as IStyleFunction<{}, {}>)({}) as IDropdownStyleProps;
         expect(styles).toMatchInlineSnapshot(`
             Object {
+              "callout": Object {
+                "boxShadow": "var(--ui-box-shadow-small)",
+              },
               "errorMessage": Array [
                 Object {
                   "backgroundColor": "var(--vscode-inputValidation-errorBackground)",