+++ packages/ui5-test-writer/src/fiori-elements-opa-writer.ts
@@ -24,6 +24,7 @@ import {
     addPagesToJourneyRunner,
     hasVirtualOPA5,
     readHtmlTargetFromQUnitJs,
+    addVirtualTestConfig,
     type JourneyRunnerPage
 } from './utils/opaQUnitUtils.js';
 import { getPackageScripts } from '@sap-ux/fiori-generator-shared';
@@ -40,6 +41,7 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
  * @param opaConfig.scriptName - the name of the OPA journey file. If not specified, 'FirstJourney' will be used
  * @param opaConfig.htmlTarget - the name of the html that will be used in OPA journey file. If not specified, 'index.html' will be used
  * @param opaConfig.appID - the appID. If not specified, will be read from the manifest in sap.app/id
+ * @param opaConfig.useVirtualPreviewEndpoints - when true, OPA harness files are served virtually; skip writing them to disk
  * @param metadata - optional metadata for the OPA test generation
  * @param fs - an optional reference to a mem-fs editor
  * @param log - optional logger instance
@@ -48,7 +50,7 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
  */
 export async function generateOPAFiles(
     basePath: string,
-    opaConfig: { scriptName?: string; appID?: string; htmlTarget?: string },
+    opaConfig: { scriptName?: string; appID?: string; htmlTarget?: string; useVirtualPreviewEndpoints?: boolean },
     metadata?: string,
     fs?: Editor,
     log?: Logger,
@@ -97,8 +99,15 @@ export async function generateOPAFiles(
             writeJourneyFiles(appFeatures, standaloneWriteContext, true, hasJourneyRunner, virtualOPA5Configured);
         }
     } else {
-        writeCommonAndPageFiles(writeContext, rootCommonTemplateDirPath);
-        writeJourneyFiles(appFeatures, writeContext, false);
+        writeCommonAndPageFiles(writeContext, rootCommonTemplateDirPath, opaConfig.useVirtualPreviewEndpoints ?? false);
+        writeJourneyFiles(appFeatures, writeContext, false, false, opaConfig.useVirtualPreviewEndpoints ?? false);
+        if (opaConfig.useVirtualPreviewEndpoints) {
+            await addVirtualTestConfig(
+                basePath,
+                [{ framework: 'OPA5', path: '/test/integration/opaTests.qunit.html' }, { framework: 'Testsuite' }],
+                editor
+            );
+        }
     }
 
     return editor;
@@ -390,21 +399,28 @@ function findLROP(
  *
  * @param writeContext - shared write context (config, paths, editor, journey params)
  * @param rootCommonTemplateDirPath - template root directory for common files
+ * @param useVirtualPreviewEndpoints - when true, testsuite harness files are served virtually; skip writing them to disk
  */
-function writeCommonAndPageFiles(writeContext: WriteContext, rootCommonTemplateDirPath: string): void {
+function writeCommonAndPageFiles(
+    writeContext: WriteContext,
+    rootCommonTemplateDirPath: string,
+    useVirtualPreviewEndpoints = false
+): void {
     const { config, rootV4TemplateDirPath, testOutDirPath, editor, journeyParams } = writeContext;
 
-    // Common test files
-    editor.copyTpl(
-        join(rootCommonTemplateDirPath),
-        testOutDirPath,
-        // unit tests are not added for Fiori elements app
-        { appId: config.appID },
-        undefined,
-        {
-            globOptions: { dot: true }
-        }
-    );
+    // Common test files (testsuite served virtually when useVirtualPreviewEndpoints is enabled)
+    if (!useVirtualPreviewEndpoints) {
+        editor.copyTpl(
+            join(rootCommonTemplateDirPath),
+            testOutDirPath,
+            // unit tests are not added for Fiori elements app
+            { appId: config.appID },
+            undefined,
+            {
+                globOptions: { dot: true }
+            }
+        );
+    }
 
     config.pages.forEach((page) => {
         writePageObject(page, rootV4TemplateDirPath, testOutDirPath, editor);