+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/project-access
 
+## 1.37.0
+
+### Minor Changes
+
+- 72695e5: chore: drop Node 20 support as it is no longer maintained
+
+### Patch Changes
+
+- Updated dependencies [72695e5]
+    - @sap-ux/i18n@0.4.0
+    - @sap-ux/ui5-config@0.31.0
+
+## 1.36.5
+
+### Patch Changes
+
+- 50a8ba5: chore: fresh release after workflow updates
+- Updated dependencies [50a8ba5]
+    - @sap-ux/i18n@0.3.12
+    - @sap-ux/ui5-config@0.30.5
+
 ## 1.36.4
 
 ### Patch Changes